refactor: 보다 명확한 이름으로 클래스명 변경(코드리뷰 반영)
